+++ docs/Changelog.md
@@ -2,6 +2,8 @@
 
 ## Next version
 
+- Color filter theme resources [#1175](https://github.com/mapsforge/vtm/pull/1175)
+  - `ThemeCallback.getBitmap`
 - Update JTS, OkHttp dependencies [#1172](https://github.com/mapsforge/vtm/pull/1172) [#1173](https://github.com/mapsforge/vtm/pull/1173)
 - Remove SLF4J dependency [#1167](https://github.com/mapsforge/vtm/pull/1167) [#1170](https://github.com/mapsforge/vtm/pull/1170)
 - Disable `vtm-ios` [#1166](https://github.com/mapsforge/vtm/pull/1166)